Deckbuilding Average

+++ ProjectDoubleSI/Assets/Scripts/DeckBuilder/CardSCO.cs
@@ -8,7 +8,7 @@ public class CardSCO : ScriptableObject
 {
     public string cardName;
     public int level;
-    public int cardCost;
+    public float cardCost;
 
     public Sprite cardAsset;
     public string cardEffet;
